refactor invocation parameter input types
